better

+++ ft_putstr.c
@@ -1,8 +1,8 @@
 #include "ft_printf.h"
 
-int	ft_putstr(char *str)
+size_t	ft_putstr(char *str)
 {
-    int i;
+    size_t i;
 
     i = 0;
 	while (str[i])
